requirements.txt
python-telegram-bot==21.6
ollama==0.3.3
aiosqlite==0.20.0
apscheduler==3.10.4
httpx==0.27.2
python-dotenv==1.0.1
tavily-python==0.5.0
